+++ gcovr/gcov_parser.py
@@ -61,7 +61,7 @@ from .coverage import (
     LineCoverage,
 )
 from .merging import (
-    MergeOptions,
+    FUNCTION_MAX_LINE_MERGE_OPTIONS,
     insert_branch_coverage,
     insert_function_coverage,
     insert_line_coverage,
@@ -349,7 +349,7 @@ def parse_coverage(
         insert_function_coverage(
             coverage,
             FunctionCoverage(name, lineno=state.lineno + 1, call_count=calls),
-            MergeOptions(ignore_function_lineno=True),
+            FUNCTION_MAX_LINE_MERGE_OPTIONS,
         )
 
     _report_lines_with_errors(
@@ -407,12 +407,12 @@ def _gather_coverage_from_line(
 
         # handle deferred functions
         for function in state.deferred_functions:
-            name, hits, _, _ = function
+            name, count, _, _ = function
 
             insert_function_coverage(
                 coverage,
-                FunctionCoverage(name, lineno=lineno, call_count=hits),
-                MergeOptions(ignore_function_lineno=True),
+                FunctionCoverage(name, lineno=lineno, call_count=count),
+                FUNCTION_MAX_LINE_MERGE_OPTIONS,
             )
 
         return _ParserState(