+++ gcovr/writer/json.py
@@ -43,6 +43,7 @@ from ..coverage import (
     sort_coverage,
 )
 from ..merging import (
+    get_merge_mode_from_options,
     insert_branch_coverage,
     insert_decision_coverage,
     insert_file_coverage,
@@ -190,14 +191,15 @@ def gcovr_json_files_to_coverage(filenames, options) -> CovData:
                 continue
 
             file_coverage = FileCoverage(file_path)
+            merge_options = get_merge_mode_from_options(options)
             for json_function in gcovr_file["functions"]:
                 insert_function_coverage(
-                    file_coverage, _function_from_json(json_function)
+                    file_coverage, _function_from_json(json_function), merge_options
                 )
             for json_line in gcovr_file["lines"]:
                 insert_line_coverage(file_coverage, _line_from_json(json_line))
 
-            insert_file_coverage(covdata, file_coverage)
+            insert_file_coverage(covdata, file_coverage, merge_options)
 
     return covdata
 
@@ -278,15 +280,20 @@ def _json_from_call(call: CallCoverage) -> dict:
 
 
 def _json_from_functions(functions: Dict[str, FunctionCoverage]) -> list:
-    return [_json_from_function(functions[name]) for name in sorted(functions)]
+    return [
+        f for name in sorted(functions) for f in _json_from_function(functions[name])
+    ]
 
 
-def _json_from_function(function: FunctionCoverage) -> dict:
-    return {
-        "name": function.name,
-        "lineno": function.lineno,
-        "execution_count": function.count,
-    }
+def _json_from_function(function: FunctionCoverage) -> list:
+    return [
+        {
+            "name": function.name,
+            "lineno": lineno,
+            "execution_count": count,
+        }
+        for lineno, count in function.count.items()
+    ]
 
 
 def _function_from_json(json_function: dict) -> FunctionCoverage: